requirements.txt
psycopg2-binary == 2.8.5
pandas == 1.1.0
openpyxl == 3.0.4
slackclient == 2.9.3
hvac == 0.10.5
